Rename EditorTypes file
